Support multi-threading in H5VL (#41)
#### Support multi-threading in the H5VL module (H5VL.c, H5VLcallback.c, H5VLdyn_ops.c, and H5VLint.c)

The global mutex is no longer acquired upon entry to H5VL API routines. The high-level outline for the changes is in the H5VL threadsafety document, but the major changes are:

- H5VL_t and H5VL_object_t are now reference counted atomically
- The dynamic optional VOL operation table is initialized atomically
- VOL Connector registration has been refactored to eliminate dangerous regions between retrieval of a VOL connector and incrementing of its reference count
- Operations within H5VL that are unsafe have been wrapped with `H5_API_LOCK`/`H5_API_UNLOCK` (Native VOL registration, routines in H5T/H5S, etc.)
- Many routines have had operations reordered and failure handling modified to eliminate dangerous regions

#### Push global lock acquisition down to VOL layer for most VOL-using API routines

For routines that bottom out in a VOL operation (Most major API routines in H5A, H5F, H5D, H5G, H5O, H5L, H5Tcommit), the global mutex is now only acquired right before VOL callback invocation, and only if the active VOL connector (or next active VOL connector in the stack) does not have the threadsafe capability flag.

#### Add dedicated tests for multithreaded H5VL behavior in testmthdf5

The individual tests are implemented in `mt_vl_test.c`.

#### Add MT Native Wrapper VOL and MT Passthru Wrapper VOL

These connectors are used to test the H5VL module's ability manage concurrent thread entry into a VOL connector/back into the VOL layer from a passthrough connector. The MT Native Wrapper is a connector with the threadsafe flag enabled, which simply grabs the global mutex before invoking the native VOL operation. The MT Passthru Wrapper is a clone of the passthrough VOL with the threadsafe flag enabled.

#### Add MT VL Test Connector

This VOL implements a few operations in a non-API compliant manner in order to test VOL connector loading on file open failure. It shouldn't be used outside of the MT VL tests.